fix(llm): use Decimal for cost_usd instead of float
LLMResult.cost_usd was annotated float but the underlying column is
NUMERIC(10, 6), which SQLAlchemy/asyncpg round-trip as Decimal. A value
read back from the database is a Decimal regardless of the annotation,
so `cost_usd + 0.05` anywhere downstream raises `TypeError: unsupported
operand type(s)` the moment real accounting arithmetic touches it —
and pyright can't catch it because the annotation says float.

Make Decimal the type end-to-end: LLMResult.cost_usd, MockLLM's return
value, and the model's Mapped annotation. Added a regression test
asserting the value is a Decimal at both the LLMResult and LLMCall
stage and that Decimal arithmetic against it doesn't raise.

+++ src/db/models.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import UTC, datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, ForeignKey, Numeric, String
 from sqlalchemy.dialects.postgresql import UUID
@@ -40,7 +41,7 @@ class LLMCall(Base):
     input_tokens: Mapped[int] = mapped_column(nullable=False)
     output_tokens: Mapped[int] = mapped_column(nullable=False)
     latency_ms: Mapped[int] = mapped_column(nullable=False)
-    cost_usd: Mapped[float] = mapped_column(Numeric(10, 6), nullable=False)
+    cost_usd: Mapped[Decimal] = mapped_column(Numeric(10, 6), nullable=False)
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), default=utcnow, nullable=False
     )
